requirements.txt

pandas
requests
openpyxl
python-dotenv
typer